Added logo.
