--- Packs/Kubernetes/Integrations/Kubernetes/Pipfile
@@ -1,18 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pylint = "*"
-pytest = "==5.0.1"
-pytest-mock = "*"
-requests-mock = "*"
-pytest-asyncio = "*"
-
-[packages]
-pytest = "*"
-requests = "*"
-
-[requires]
-python_version = "3.7"
