+++ Packs/Kubernetes/pack_metadata.json
@@ -2,10 +2,10 @@
     "name": "Kubernetes",
     "description": "Manages Kubernetes clusters.",
     "support": "community",
-    "currentVersion": "1.0.0",
+    "currentVersion": "0.1.0",
     "author": "Allister Beharry",
     "url": "https://live.paloaltonetworks.com/t5/cortex-xsoar-discussions/bd-p/Cortex_XSOAR_Discussions",
-    "email": "",
+    "email": "allister.beharry@gmail.com",
     "created": "2020-09-22T17:16:26Z",
     "categories": [
         "IT Services"
@@ -13,5 +13,5 @@
     "tags": [],
     "useCases": [],
     "keywords": [],
-    "githubUser": []
+    "githubUser": ["allisterb"]
 }
