--- Pipfile
@@ -1,12 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-pytest-parametrized = "*"
-
-[requires]
-python_version = "3.8"
